RR-439 - Refactored induction set mapping into a mapper class in it's own right
