+++ server/data/mappers/otherQualificationsMapper.ts
@@ -1,5 +1,6 @@
 import type { CiagInduction } from 'ciagInductionApiClient'
 import type { OtherQualifications } from 'viewModels'
+import toInductionQuestionSet from './inductionQuestionSetMapper'
 
 const toOtherQualifications = (ciagInduction: CiagInduction): OtherQualifications => {
   if (!ciagInduction || !ciagInduction.qualificationsAndTraining) {
@@ -17,15 +18,6 @@ const toOtherQualifications = (ciagInduction: CiagInduction): OtherQualification
     highestEducationLevel: ciagInduction.qualificationsAndTraining.educationLevel,
     additionalTraining: ciagInduction.qualificationsAndTraining.additionalTraining,
   }
-}
-
-const toInductionQuestionSet = (
-  ciagInduction: CiagInduction,
-): 'LONG_QUESTION_SET' | 'SHORT_QUESTION_SET' | undefined => {
-  if (ciagInduction) {
-    return ciagInduction.hopingToGetWork === 'YES' ? 'LONG_QUESTION_SET' : 'SHORT_QUESTION_SET'
-  }
-  return undefined
 }
 
 export default toOtherQualifications
