+++ server/data/mappers/workAndInterestMapper.ts
@@ -7,6 +7,7 @@ import type {
   WorkExperience,
   WorkInterests,
 } from 'viewModels'
+import toInductionQuestionSet from './inductionQuestionSetMapper'
 
 const toWorkAndInterests = (ciagInduction: CiagInduction): WorkAndInterests => {
   return {
@@ -79,15 +80,6 @@ const toWorkInterests = (ciagInduction: CiagInduction): WorkInterests => {
     updatedBy: ciagInduction.workExperience.workInterests.modifiedBy,
     updatedAt: moment(ciagInduction.workExperience.workInterests.modifiedDateTime).toDate(),
   }
-}
-
-const toInductionQuestionSet = (
-  ciagInduction: CiagInduction,
-): 'LONG_QUESTION_SET' | 'SHORT_QUESTION_SET' | undefined => {
-  if (ciagInduction) {
-    return ciagInduction.hopingToGetWork === 'YES' ? 'LONG_QUESTION_SET' : 'SHORT_QUESTION_SET'
-  }
-  return undefined
 }
 
 export default toWorkAndInterests
